#1b Add class: Place class

+++ app/src/main/java/team2/kakigowherebackend/model/Tourist.java
@@ -11,9 +11,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Tourist {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int touristId;
+    private long touristId;
     private String touristEmail;
     private String touristPassword;
     private String touristName;
